Added the Raspbian OS /boot/movefs script
This script removes unused packages from the initial Rasbian OS,
upgrades the OS, and then moves the RootFS off the SD Card and onto a
USB device that is labeled as /dev/sda.  To ensure this works correctly,
you need to be logged in as the "root" user, and only one USB storage
device can be attached to the RPi.